Consolidate diagnosis service implementations
