+++ api/src/test/java/org/openmrs/module/emrapi/visit/VisitDomainWrapperComponentTest.java
@@ -72,6 +72,7 @@ public class VisitDomainWrapperComponentTest extends BaseModuleContextSensitiveT
         dispositionDescriptor = ContextSensitiveMetadataTestUtils.setupDispositionDescriptor(conceptService, dispositionService);
         ContextSensitiveMetadataTestUtils.setupAdmissionDecisionConcept(conceptService, emrApiProperties);
         ContextSensitiveMetadataTestUtils.setupDiagnosisMetadata(conceptService, emrApiProperties);
+        Context.getAdministrationService().setGlobalProperty(EmrApiConstants.GP_USE_LEGACY_DIAGNOSIS_SERVICE, "true");
     }
 
     @Test
